+++ src/main/java/laddergame/exception/LadderGameException.java
@@ -1,4 +1,6 @@
-package laddergame;
+package laddergame.exception;
+
+import laddergame.exception.ErrorCode;
 
 public class LadderGameException extends RuntimeException {
     public LadderGameException(ErrorCode errorCode) {
