+++ src/test/java/laddergame/domain/PlayersTest.java
@@ -1,5 +1,9 @@
-package laddergame;
+package laddergame.domain;
 
+import laddergame.domain.Player;
+import laddergame.domain.Players;
+import laddergame.exception.ErrorCode;
+import laddergame.exception.LadderGameException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
